- Add installability check (--installable / --later-installable)
The strategy is to not use a solver (so we need to do a fixpoint computation).
We look at each package individually, and figure out whether their dependencies
can be fulfilled by the opam-repository (we treat filters with "dev"/"with-doc"/
"with-test"/"with-dev-setup" as always fulfilled).
For each dependency we look up in our opam-repository whether the package name
is present and has candidates, and use OpamVersionCompare.compare we manually
solve the dependencies. This is of course incomplete/not exact. That's fine ;)

The advantage over "opam admin check -i --installability" is that we consider
e.g. all "os" filters. We also remember which constraint wasn't satisfied and
provide this as reasoning.

Now, the fixpoint is a bit tricky: due to dry-run we keep track of (a) opam
files that we moved (or would like to move, but are not allowed) "moved_files",
(b) the packages provided by `--pkg` (or `--pkg-file`) "pkgs", and (c) our
opam repository "opams" (initialized to
"OpamRepositoryState.load_opams_from_dir .. <opam-repository>").

So, once we ran the filter (and if the filter is installable, or
--later-installable is provided), we update "pkgs" and "opams" by removing those
elements that we found in "moved_files", and in case we added something to
"moved_files", we execute the filter again. Note that we skip opam files present
in "moved_files" (so we don't accidentally try to remove the same file twice -
joking, it is for the dry-run again).

The other reason for this installability in contrast to "opam admin check" is
the runtime - on my machine, opam admin check takes 4 hours, and one round of
opam-archive --installable --pkg-all --dry-run takes 10 minutes (now, in the
case for unavailable, phase 1, it needs 4 iterations - so it is 40 minutes).

The great advantage is as well that we retrieve the explanation (the offending
constraint), which may be useful for someone who wants to fix these constraints.

- Also flip flag 'summary' to 'no-summary'

- Add '--pkg-file <filename>' and '--pkg-all' options to provide packages via
  a file or considering all packages (useful for --installable and
  --later-installable).

- Refactor code, split it into more readable chunks (ehm, functions).